extract object setup code into helper and make tests more readable
