+++ app/test/unit/stores/updates/update-conflict-state-test.ts
@@ -1,23 +1,5 @@
 import { updateConflictState } from '../../../../src/lib/stores/updates/changes-state'
-import { IStatusResult } from '../../../../src/lib/git'
-import { IChangesState } from '../../../../src/lib/app-state'
-import { WorkingDirectoryStatus } from '../../../../src/models/status'
-
-const baseChangesState: IChangesState = {
-  workingDirectory: WorkingDirectoryStatus.fromFiles([]),
-  selectedFileIDs: [],
-  diff: null,
-  commitMessage: null,
-  showCoAuthoredBy: false,
-  coAuthors: [],
-  conflictState: null,
-}
-
-const baseStatus: IStatusResult = {
-  exists: true,
-  mergeHeadFound: false,
-  workingDirectory: WorkingDirectoryStatus.fromFiles([]),
-}
+import { createState, createStatus } from './changes-state-helper'
 
 describe('updateConflictState', () => {
   const statsStore = {
@@ -26,35 +8,33 @@ describe('updateConflictState', () => {
   }
 
   it('returns null when no MERGE_HEAD file found', () => {
-    const prevState = { ...baseChangesState, conflictState: null }
-    const status = { ...baseStatus, mergeHeadFound: false }
+    const prevState = createState({ conflictState: null })
+    const status = createStatus({ mergeHeadFound: false })
     const { conflictState } = updateConflictState(status, statsStore, prevState)
     expect(conflictState).toBeNull()
   })
 
   it('returns null when MERGE_HEAD set but not branch or tip defined', () => {
-    const prevState = { ...baseChangesState, conflictState: null }
-    const status = {
-      ...baseStatus,
+    const prevState = createState({ conflictState: null })
+    const status = createStatus({
       mergeHeadFound: true,
       currentBranch: undefined,
       currentTip: undefined,
-    }
+    })
+
     const { conflictState } = updateConflictState(status, statsStore, prevState)
     expect(conflictState).toBeNull()
   })
 
   it('returns a value when status has MERGE_HEAD set', () => {
-    const prevState = {
-      ...baseChangesState,
+    const prevState = createState({
       conflictState: null,
-    }
-    const status = {
-      ...baseStatus,
+    })
+    const status = createStatus({
       mergeHeadFound: true,
       currentBranch: 'master',
       currentTip: 'first-sha',
-    }
+    })
 
     const { conflictState } = updateConflictState(status, statsStore, prevState)
 
@@ -65,19 +45,17 @@ describe('updateConflictState', () => {
   })
 
   it('increments abort counter when branch has changed', () => {
-    const prevState = {
-      ...baseChangesState,
+    const prevState = createState({
       conflictState: {
         currentBranch: 'old-branch',
         currentTip: 'old-sha',
       },
-    }
-    const status = {
-      ...baseStatus,
+    })
+    const status = createStatus({
       mergeHeadFound: true,
       currentBranch: 'master',
       currentTip: 'first-sha',
-    }
+    })
 
     updateConflictState(status, statsStore, prevState)
 
@@ -85,19 +63,17 @@ describe('updateConflictState', () => {
   })
 
   it('increments abort counter when conflict resolved and tip has not changed', () => {
-    const prevState = {
-      ...baseChangesState,
+    const prevState = createState({
       conflictState: {
         currentBranch: 'master',
         currentTip: 'old-sha',
       },
-    }
-    const status = {
-      ...baseStatus,
+    })
+    const status = createStatus({
       mergeHeadFound: false,
       currentBranch: 'master',
       currentTip: 'old-sha',
-    }
+    })
 
     updateConflictState(status, statsStore, prevState)
 
@@ -105,19 +81,17 @@ describe('updateConflictState', () => {
   })
 
   it('increments success counter when conflict resolved and tip has changed', () => {
-    const prevState = {
-      ...baseChangesState,
+    const prevState = createState({
       conflictState: {
         currentBranch: 'master',
         currentTip: 'old-sha',
       },
-    }
-    const status = {
-      ...baseStatus,
+    })
+    const status = createStatus({
       mergeHeadFound: false,
       currentBranch: 'master',
       currentTip: 'new-sha',
-    }
+    })
 
     updateConflictState(status, statsStore, prevState)
